Use the candidate IntegerBuffer in the live H1 bridge
Align the live owner bridge and final-driver helper with the producer's packed IntegerBuffer preparation state. Pass that owner directly through the unified native root and remove the redundant eight-word Int64Buffer allocation and copy.

Update native checkers and worker allocation metadata to enforce the exact ABI and retain cross-backend and unified end-to-end coverage.

+++ bench/pari-class-group-port/class_group_final_driver_status.py
@@ -33,7 +33,7 @@ class FinalDriverStatusFailure(ValueError):
 @native
 def pari_equal_bound_cleanarch_driver_status(
     prep_base_state: IntegerBuffer,
-    prep_state: Int64Buffer,
+    prep_state: IntegerBuffer,
     hnf_state: Int64Buffer,
     acceptance_state: Int64Buffer,
     class_logs: IntegerBuffer,
